moved computedelaymatrix function to /utils/model_utils
