+++ neurolib/models/ww/timeIntegration.py
@@ -1,7 +1,7 @@
 import numpy as np
 import numba
 
-from . import loadDefaultParams as dp
+from ...utils import model_utils as mu
 
 
 def timeIntegration(params):
@@ -63,7 +63,7 @@ def timeIntegration(params):
         Dmat = np.zeros((N, N))
     else:
         # Interareal connection delays, Dmat(i,j) Connnection from jth node to ith (ms)
-        Dmat = dp.computeDelayMatrix(lengthMat, signalV)
+        Dmat = mu.computeDelayMatrix(lengthMat, signalV)
         # no self-feedback delay
         Dmat[np.eye(len(Dmat)) == 1] = np.zeros(len(Dmat))
     Dmat_ndt = np.around(Dmat / dt).astype(int)  # delay matrix in multiples of dt
