+++ neurolib/utils/model_utils.py
@@ -42,3 +42,24 @@ def adjustArrayShape(original, target):
     original = original[: target.shape[0], -target.shape[1] :]
 
     return original
+
+
+def computeDelayMatrix(lengthMat, signalV, segmentLength=1):
+    """
+    Compute the delay matrix from the fiber length matrix and the signal
+    velocity
+
+        :param lengthMat:       A matrix containing the connection length in
+            segment
+        :param signalV:         Signal velocity in m/s
+        :param segmentLength:   Length of a single segment in mm
+
+        :returns:    A matrix of connexion delay in ms
+    """
+
+    normalizedLenMat = lengthMat * segmentLength
+    if signalV > 0:
+        Dmat = normalizedLenMat / signalV  # Interareal delays in ms
+    else:
+        Dmat = lengthMat * 0.0
+    return Dmat
